fix: another fix is parsing batch
* two or more MSH in the Batch function should work
* if one MSH is passed in Batch, fail
* moved unit tests for sanity checks to new file
* updated unit tests
* using HL&FatalError class for issues inside normalizedBuilder.ts and normalizedClient.ts
* added jest verbose as an option
