+++ __tests__/__data__/constants.ts
@@ -1,0 +1,7 @@
+import { HL7_2_7_MSH } from '../../src/specification/2.7'
+
+export const MSH_HEADER: HL7_2_7_MSH = {
+  msh_9_1: 'ADT',
+  msh_9_2: 'A01',
+  msh_11_1: 'D'
+}
