+++ src/utils/normalizedBuilder.ts
@@ -1,7 +1,9 @@
 import fs from 'fs'
 import { HL7_2_7 } from '../specification/2.7.js'
 import { MSH } from '../specification/specification'
+import { HL7FatalError } from './exception.js'
 import { ParserPlan } from './parserPlan.js'
+import { isBatch } from './utils.js'
 
 const DEFAULT_CLIENT_BUILDER_OPTS = {
   date: '14',
@@ -111,13 +113,13 @@ export function normalizedClientMessageBuilderOptions (raw?: ClientBuilderMessag
   const props: ClientBuilderMessageOptions = { ...DEFAULT_CLIENT_BUILDER_OPTS, ...raw }
 
   if (typeof props.messageHeader === 'undefined' && props.text === '') {
-    throw new Error('mshHeader must be set if no HL7 message is being passed.')
+    throw new HL7FatalError(500, 'mshHeader must be set if no HL7 message is being passed.')
   } else if (typeof props.messageHeader === 'undefined' && typeof props.text !== 'undefined' && props.text.slice(0, 3) !== 'MSH') {
     throw new Error('text must begin with the MSH segment.')
   }
 
   if ((typeof props.newLine !== 'undefined' && props.newLine === '\\r') || props.newLine === '\\n') {
-    throw new Error('newLine must be \r or \n')
+    throw new HL7FatalError(500, ('newLine must be \r or \n'))
   }
 
   if (props.date !== '8' && props.date !== '12' && props.date !== '14') {
@@ -147,11 +149,15 @@ export function normalizedClientBatchBuilderOptions (raw?: ClientBuilderOptions)
   const props: ClientBuilderOptions = { ...DEFAULT_CLIENT_BUILDER_OPTS, ...raw }
 
   if (typeof props.text !== 'undefined' && props.text !== '' && props.text.slice(0, 3) !== 'BHS' && props.text.slice(0, 3) !== 'MSH') {
-    throw new Error('text must begin with the BHS or MSH segment.')
+    throw new HL7FatalError(500, ('text must begin with the BHS or MSH segment.'))
+  }
+
+  if (typeof props.text !== 'undefined' && props.text !== '' && props.text.slice(0, 3) === 'MSH' && !isBatch(props.text)) {
+    throw new HL7FatalError(500, ('Unable to process a single MSH as a batch. Use Message.'))
   }
 
   if ((typeof props.newLine !== 'undefined' && props.newLine === '\\r') || props.newLine === '\\n') {
-    throw new Error('newLine must be \r or \n')
+    throw new HL7FatalError(500, ('newLine must be \r or \n'))
   }
 
   if (props.date !== '8' && props.date !== '12' && props.date !== '14') {
@@ -179,19 +185,19 @@ export function normalizedClientFileBuilderOptions (raw?: ClientBuilderFileOptio
   const props: ClientBuilderFileOptions = { ...DEFAULT_CLIENT_FILE_OPTS, ...DEFAULT_CLIENT_BUILDER_OPTS, ...raw }
 
   if (typeof props.text !== 'undefined' && props.text !== '' && props.text.slice(0, 3) !== 'FHS') {
-    throw new Error('text must begin with the FHS segment.')
+    throw new HL7FatalError(500, ('text must begin with the FHS segment.'))
   }
 
   if ((typeof props.newLine !== 'undefined' && props.newLine === '\\r') || props.newLine === '\\n') {
-    throw new Error('newLine must be \r or \n')
+    throw new HL7FatalError(500, ('newLine must be \r or \n'))
   }
 
   if (typeof props.extension !== 'undefined' && props.extension.length !== 3) {
-    throw new Error('The extension for file save must be 3 characters long.')
+    throw new HL7FatalError(500, ('The extension for file save must be 3 characters long.'))
   }
 
   if (typeof props.fullFilePath !== 'undefined' && typeof props.fileBuffer !== 'undefined') {
-    throw new Error('You can not have specified a file path and a buffer. Please choose one or the other.')
+    throw new HL7FatalError(500, ('You can not have specified a file path and a buffer. Please choose one or the other.'))
   }
 
   if (props.date !== '8' && props.date !== '12' && props.date !== '14') {
