+++ src/utils/normalizedClient.ts
@@ -1,6 +1,7 @@
 import { TcpSocketConnectOpts } from 'node:net'
 import type { ConnectionOptions as TLSOptions } from 'node:tls'
 import { InboundResponse } from '../client/module/inboundResponse.js'
+import { HL7FatalError } from './exception.js'
 import { assertNumber, validIPv4, validIPv6 } from './utils.js'
 
 /**
@@ -124,22 +125,22 @@ export function normalizeClientOptions (raw?: ClientOptions): ValidatedClientOpt
   const props: any = { ...DEFAULT_CLIENT_OPTS, ...raw }
 
   if (typeof props.host === 'undefined' || props.host.length <= 0) {
-    throw new Error('host is not defined or the length is less than 0.')
+    throw new HL7FatalError(500, 'host is not defined or the length is less than 0.')
   }
 
   if (props.ipv4 === true && props.ipv6 === true) {
-    throw new Error('ipv4 and ipv6 both can\'t be set to be both used exclusively.')
+    throw new HL7FatalError(500, 'ipv4 and ipv6 both can\'t be set to be both used exclusively.')
   }
 
   if (typeof props.host !== 'string' && props.ipv4 === false && props.ipv6 === false) {
-    throw new Error('host is not valid string.')
+    throw new HL7FatalError(500, 'host is not valid string.')
   } else if (typeof props.host === 'string' && props.ipv4 === true && props.ipv6 === false) {
     if (!validIPv4(props.host)) {
-      throw new Error('host is not a valid IPv4 address.')
+      throw new HL7FatalError(500, 'host is not a valid IPv4 address.')
     }
   } else if (typeof props.host === 'string' && props.ipv4 === false && props.ipv6 === true) {
     if (!validIPv6(props.host)) {
-      throw new Error('host is not a valid IPv6 address.')
+      throw new HL7FatalError(500, 'host is not a valid IPv6 address.')
     }
   }
 
@@ -158,11 +159,11 @@ export function normalizeClientListenerOptions (raw?: ClientListenerOptions): Va
   const props: any = { ...DEFAULT_LISTEN_CLIENT_OPTS, ...raw }
 
   if (typeof props.port === 'undefined') {
-    throw new Error('port is not defined.')
+    throw new HL7FatalError(500, 'port is not defined.')
   }
 
   if (typeof props.port !== 'number') {
-    throw new Error('port is not valid number.')
+    throw new HL7FatalError(500, 'port is not valid number.')
   }
 
   assertNumber(props, 'connectionTimeout', 0)
